shopfloor_mobile_base: fix scenario state load
The init state is called by the  bedoreRouteUpdate method of the router. At this stage, the new route is not yet set into the contexte and therefore, the menu item id is still the one from the previous route. To ensure that the new menu id is properly set before the init state is entered, we force the menu id to the one of the new route. Call to odoo into the enter method of the init state will be made with the right menu id into the request headers.

fixes #359