Fix shopfloor edit picking carrier form
There is an issue with the get url on the form although there is two url
options one with `/get` at the end and one without. Only the former works.
The one not working returns a 405.
This is why the call is changed in the js file.

Another problem for both url is their route are recorded with <int:id>
instead of <int:_id> wich is the default for a get route to work with
the `base_rest` module.
The workaround for this is to add a validator for the `id` parameter.